Fix to #8862 - The multi-part identifier "ts.Id" could not be bound
When we try to bind to parent query model we keep track of the current model output data info. However when recursive binding is required (i.e. binding to grandparent) we were not storing the output data of the parent, but still using the output data info of the child. This caused problems when the parent output data info doesn't allow for the binding to its parent - we would still try to perform the binding because we took the incorrect output data info into account.

Fix is to store the correct output data info in the variable before recursively trying to call TryBindParentExpression.

(cherry picked from commit c6f13a003759bd8b88e678d4dc77c10b7c46bf26)