Fix to #9395 - Query: Apply CompositePredicateExpressionVisitor to join predicates
Some optimizations that are applied to predicates in SelectExpression can be safely applied to join predicates as well. Fix modifies CompositePredicateExpressionVisitor to also look into JoinExpression predicates and not only SelectExpression predicates.
Also changes the way we build join predicates themselves, when inheritance is concerned - now the discriminator predicate will be extracted out of the comparison, so that the comparisons are sargable.

Before:

((a is Derived) ? a.DerivedFK : null) == b.PK - which translated to CASE block (not sargable)

After:

a is Derived && a.DerivedFK == b.PK